+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityConditionFieldFactory.java
@@ -25,9 +25,9 @@ import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.attribute.Attribute;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
-import is.codion.framework.model.AbstractForeignKeyConditionModel;
 import is.codion.framework.model.DefaultForeignKeyConditionModel;
 import is.codion.framework.model.EntitySearchModel;
+import is.codion.framework.model.ForeignKeyConditionModel;
 import is.codion.swing.common.ui.component.combobox.Completion;
 import is.codion.swing.common.ui.component.table.FilterColumnConditionPanel.FieldFactory;
 import is.codion.swing.common.ui.component.value.ComponentValue;
@@ -139,8 +139,8 @@ public final class EntityConditionFieldFactory implements FieldFactory<Attribute
 	}
 
 	private JComponent createInForeignKeyField(ColumnConditionModel<? extends Attribute<?>, ?> model) {
-		if (model instanceof AbstractForeignKeyConditionModel) {
-			EntitySearchModel searchModel = ((AbstractForeignKeyConditionModel) model).inSearchModel();
+		if (model instanceof ForeignKeyConditionModel) {
+			EntitySearchModel searchModel = ((ForeignKeyConditionModel) model).inSearchModel();
 
 			return configureSearchField(searchModel, inputComponents
 							.foreignKeySearchField((ForeignKey) model.columnIdentifier(), searchModel).build());